move zip.py module to utils directory, re #5582
